chore: 디렉토리명 ui -> resolver로 수정
