+++ src/main/java/com/interview_master/resolver/QuizAttemptsResolver.java
@@ -1,9 +1,10 @@
-package com.interview_master.ui;
+package com.interview_master.resolver;
 
 import static com.interview_master.util.GraphQLAuthUtils.validateUserAuthContext;
 
-import com.interview_master.dto.QuizWithAttempt;
-import com.interview_master.infrastructure.QuizRepository;
+import com.interview_master.service.UserQuizAttemptService;
+import com.interview_master.dto.QuizGarden;
+import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.graphql.data.method.annotation.Argument;
@@ -13,20 +14,17 @@ import org.springframework.stereotype.Controller;
 
 @Controller
 @RequiredArgsConstructor
-public class QuizWithAttemptsResolver {
+public class QuizAttemptsResolver {
 
-  private final QuizRepository quizRepository;
+  private final UserQuizAttemptService userQuizAttemptService;
 
-  /**
-   * collection에 속한 퀴즈 목록과 각 퀴즈에 대한 현재 유저가 시도한 정보들
-   */
   @QueryMapping
-  public List<QuizWithAttempt> getQuizzesWithAttemptByCollectionId(@Argument Long collectionId,
+  public List<QuizGarden> getQuizGarden(@Argument LocalDate startDate, @Argument LocalDate endDate,
       @ContextValue(required = false) Long userId,
       @ContextValue(name = "authError", required = false) String authError) {
+
     validateUserAuthContext(userId, authError);
 
-    return quizRepository.getQuizzesByCollectionIdWithAttempts(collectionId, userId);
+    return userQuizAttemptService.getQuizGardens(startDate, endDate, userId);
   }
-
 }
