+++ src/main/java/com/interview_master/resolver/SearchCollectionResolver.java
@@ -1,11 +1,11 @@
-package com.interview_master.ui;
+package com.interview_master.resolver;
 
 import com.interview_master.dto.CollectionWithAttempt;
 import com.interview_master.dto.DataPage;
 import com.interview_master.dto.PageInfo;
 import com.interview_master.dto.SortOrder;
 import com.interview_master.service.SearchCollectionService;
-import com.interview_master.ui.response.CollectionWithAttemptsPaging;
+import com.interview_master.resolver.response.CollectionWithAttemptsPaging;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.graphql.data.method.annotation.Argument;
