+++ src/main/java/com/interview_master/resolver/request/EditQuizInput.java
@@ -1,4 +1,4 @@
-package com.interview_master.ui.request;
+package com.interview_master.resolver.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class CreateQuizInput {
+public class EditQuizInput {
 
   private String question;
