+++ src/main/java/com/interview_master/service/UpsertCollectionService.java
@@ -8,8 +8,8 @@ import com.interview_master.domain.user.User;
 import com.interview_master.infrastructure.CategoryRepository;
 import com.interview_master.infrastructure.CollectionRepository;
 import com.interview_master.infrastructure.UserRepository;
-import com.interview_master.ui.request.CreateCollectionReq;
-import com.interview_master.ui.request.EditCollectionReq;
+import com.interview_master.resolver.request.CreateCollectionReq;
+import com.interview_master.resolver.request.EditCollectionReq;
 import com.interview_master.util.ExtractUserId;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
