+++ src/main/java/com/interview_master/service/UpsertQuizService.java
@@ -8,8 +8,8 @@ import com.interview_master.domain.user.User;
 import com.interview_master.infrastructure.CollectionRepository;
 import com.interview_master.infrastructure.QuizRepository;
 import com.interview_master.infrastructure.UserRepository;
-import com.interview_master.ui.request.CreateQuizInput;
-import com.interview_master.ui.request.EditQuizInput;
+import com.interview_master.resolver.request.CreateQuizInput;
+import com.interview_master.resolver.request.EditQuizInput;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
